requirements.txt
dnspython
tqdm